Fix --get-archive=arc not waiting for ACK
After the DMPAFT command has been sent and the ACK has been
received, the driver was sending the date and was waiting for the ACK
only for a few CPU cycles, not enough for the davis to process
the request. A separate ReadBufferWithTimeout function was added
to block the execution and wait for the incomming bytes before
timing out.